#364  major upgrade for whole app no-login experience, details page loads in drawer or full page view, hoist chain and user context to AppHarnessMX, clean up context hooks, RecentDataNFTs component, landing page upgrade, new AppHeader Component, fix to path to menu enum reload logic, DataNFTDetails component chain context race condition fixed, new route for /home, AppMultiversX simplified to remove local context issues
