+++ src/AdvertiseData/SellDataMultiversX.tsx
@@ -49,14 +49,13 @@ import {
   useDisclosure,
   useToast,
   Wrap,
+  useColorMode
 } from "@chakra-ui/react";
 import { ResultsParser } from "@multiversx/sdk-core";
 import { useGetPendingTransactions, useTrackTransactionStatus } from "@multiversx/sdk-dapp/hooks";
 import { useGetAccountInfo } from "@multiversx/sdk-dapp/hooks/account";
-import axios from "axios";
-import mime from "mime";
 import { File, NFTStorage } from "nft.storage";
-import { convertWeiToEsdt, isValidNumericCharacter, MENU, sleep } from "libs/util";
+import { convertWeiToEsdt, isValidNumericCharacter, MENU, sleep, styleStrings } from "libs/util";
 import { checkBalance } from "MultiversX/api";
 import { DataNftMintContract } from "MultiversX/dataNftMint";
 import { UserDataType } from "MultiversX/types";
@@ -130,7 +129,6 @@ function makeRequest(url: string): Promise<{ statusCode: number; isError: boolea
 
 const checkUrlReturns200 = async (url: string) => {
   const { statusCode, isError } = await makeRequest(url);
-  console.log("statusCode", statusCode);
 
   let isSuccess = false;
   let message = "";
@@ -151,11 +149,11 @@ const checkUrlReturns200 = async (url: string) => {
 };
 
 export default function SellDataMX({ onRfMount, itheumAccount }: { onRfMount: any; itheumAccount: any }) {
+  const { colorMode } = useColorMode();
   const { address: mxAddress } = useGetAccountInfo();
   const { hasPendingTransactions } = useGetPendingTransactions();
   const { chainMeta: _chainMeta, setChainMeta } = useChainMeta();
   const toast = useToast();
-  // const [sellerData, setSellerData] = useState("");
   const [saveProgress, setSaveProgress] = useState({
     s1: 0,
     s2: 0,
@@ -204,8 +202,11 @@ export default function SellDataMX({ onRfMount, itheumAccount }: { onRfMount: an
   const [selectedProgramId, setSelectedProgramId] = useState(null);
 
   const mxDataNftMintContract = new DataNftMintContract(_chainMeta.networkId);
+
   // query settings from Data NFT Minter SC
   useEffect(() => {
+    // console.log('********** SellDataMultiversX LOAD _chainMeta ', _chainMeta);
+
     (async () => {
       const interaction = mxDataNftMintContract.contract.methods.getMinRoyalties();
       const query = interaction.check().buildQuery();
@@ -272,6 +273,7 @@ export default function SellDataMX({ onRfMount, itheumAccount }: { onRfMount: an
       setUserData(_userData);
     }
   };
+
   useEffect(() => {
     getUserData();
   }, [mxAddress, hasPendingTransactions]);
@@ -766,15 +768,30 @@ export default function SellDataMX({ onRfMount, itheumAccount }: { onRfMount: an
     }
   };
 
+  let gradientBorder = styleStrings.gradientBorderPassive;
+
+  if (colorMode === "light") {
+    gradientBorder = styleStrings.gradientBorderPassiveLight;
+  }
+
   return (
-    <Stack spacing={5}>
+    <Stack>
       <Heading size="lg">Trade Data</Heading>
       <Heading size="xs" opacity=".7">
         Connect, mint and trade your datasets as Data NFTs in our Data NFT Marketplace
       </Heading>
 
       <Wrap shouldWrapChildren={true} spacing={5}>
-        <Box maxW="xs" borderWidth="1px" borderRadius="lg" overflow="hidden" mt={5}>
+        <Box
+          maxW="xs"
+          borderWidth="1px"
+          overflow="hidden"
+          mt={5}
+          border=".1rem solid transparent"
+          backgroundColor="none"
+          borderRadius="1.5rem"
+          style={{ "background": gradientBorder }}>
+
           <Image src="https://itheum-static.s3.ap-southeast-2.amazonaws.com/data-stream.png" alt="" />
 
           <Box p="6">
@@ -797,7 +814,16 @@ export default function SellDataMX({ onRfMount, itheumAccount }: { onRfMount: an
           </Heading>
           <Wrap shouldWrapChildren={true} spacing={5}>
             {itheumAccount.programsAllocation.map((item: any) => (
-              <Box key={item.program} maxW="xs" borderWidth="1px" borderRadius="lg" overflow="hidden">
+              <Box
+                key={item.program}
+                maxW="xs"
+                borderWidth="1px"
+                overflow="hidden"
+                border=".1rem solid transparent"
+                backgroundColor="none"
+                borderRadius="1.5rem"
+                style={{ "background": gradientBorder }}>
+
                 <Image src={`https://itheum-static.s3-ap-southeast-2.amazonaws.com/dex-${itheumAccount._lookups.programs[item.program].img}.png`} alt="" />
 
                 <Box p="6">
