+++ src/AuthPicker/AuthPickerMultiversX.tsx
@@ -92,22 +92,22 @@ function AuthPickerMx({ launchEnvironment, resetLaunchMode }: { launchEnvironmen
                     <Wrap spacing="20px" justify="space-between" padding="10px">
                       <WrapItem onClick={() => goMxLogin(WALLETS.MX_XPORTALAPP)} className="auth_wrap">
                         <WalletConnectLoginButton
-                          callbackRoute={"/"}
+                          callbackRoute={"/home"}
                           loginButtonText={"xPortal App"}
                           buttonClassName="auth_button"
                           {...(walletConnectV2ProjectId ? { isWalletConnectV2: true } : {})}></WalletConnectLoginButton>
                       </WrapItem>
 
                       <WrapItem onClick={() => goMxLogin(WALLETS.MX_DEFI)} className="auth_wrap">
-                        <ExtensionLoginButton callbackRoute={"/"} loginButtonText={"DeFi Wallet"} buttonClassName="auth_button"></ExtensionLoginButton>
+                        <ExtensionLoginButton callbackRoute={"/home"} loginButtonText={"DeFi Wallet"} buttonClassName="auth_button"></ExtensionLoginButton>
                       </WrapItem>
 
                       <WrapItem onClick={() => goMxLogin(WALLETS.MX_WEBWALLET)} className="auth_wrap">
-                        <WebWalletLoginButton callbackRoute={"/"} loginButtonText={"Web Wallet"} buttonClassName="auth_button"></WebWalletLoginButton>
+                        <WebWalletLoginButton callbackRoute={"/home"} loginButtonText={"Web Wallet"} buttonClassName="auth_button"></WebWalletLoginButton>
                       </WrapItem>
 
                       <WrapItem onClick={() => goMxLogin(WALLETS.MX_LEDGER)} className="auth_wrap">
-                        <LedgerLoginButton callbackRoute={"/"} loginButtonText={"Ledger"} buttonClassName="auth_button"></LedgerLoginButton>
+                        <LedgerLoginButton callbackRoute={"/home"} loginButtonText={"Ledger"} buttonClassName="auth_button"></LedgerLoginButton>
                       </WrapItem>
                     </Wrap>
                   </Stack>
