+++ src/DataNFT/DataNFTDetails.tsx
@@ -16,24 +16,40 @@ type DataNFTDetailsProps = {
   owner?: string;
   listed?: number;
   showConnectWallet?: boolean;
+  tokenIdProp?: string;
+  offerId?: number;
 };
 
 export default function DataNFTDetails(props: DataNFTDetailsProps) {
   const { chainMeta: _chainMeta } = useChainMeta();
-  const { tokenId } = useParams();
+  const { tokenId: tokenIdParam } = useParams();
   const [nftData, setNftData] = useState<any>({});
   const [isLoadingDetails, setIsLoadingDetails] = useState<boolean>(true);
   const [isLoadingPrice, setIsLoadingPrice] = useState<boolean>(true);
   const navigate = useNavigate();
-  const explorerUrl = getExplorer(_chainMeta.networkId);
-  const nftExplorerUrl = getNftLink(_chainMeta.networkId, tokenId || "");
   const [price, setPrice] = useState<number>(0);
   const owner = props.owner || "";
   const listed = props.listed || 0;
   const showConnectWallet = props.showConnectWallet || false;
   const toast = useToast();
+  const tokenId = props.tokenIdProp || tokenIdParam; // priority 1 is tokenIdProp
+  let explorerUrl = '';
+  let nftExplorerUrl = '';
 
   useEffect(() => {
+    if (_chainMeta?.networkId) {
+      // console.log('********** DataNFTDetails LOAD A _chainMeta READY ', _chainMeta);
+
+      explorerUrl = getExplorer(_chainMeta.networkId);
+      nftExplorerUrl = getNftLink(_chainMeta.networkId, tokenId || "");
+
+      getTokenDetails();
+      getTokenHistory();
+
+    }
+  }, [_chainMeta]);
+
+  function getTokenDetails() {
     const apiLink = getApi(_chainMeta.networkId);
     const nftApiLink = `https://${apiLink}/nfts/${tokenId}`;
     axios.get(nftApiLink).then((res) => {
@@ -52,11 +68,11 @@ export default function DataNFTDetails(props: DataNFTDetailsProps) {
         isClosable: true,
       });
     });
-  }, []);
+  }
 
-  useEffect(() => {
+  function getTokenHistory() {
     const apiUrl = getApi(_chainMeta.networkId);
-    axios.get(`https://${apiUrl}/nfts/${tokenId}/transactions?status=success&function=addOffer&size=1&receiver=${_chainMeta.contracts.market}`).then((res) => {
+    axios.get(`https://${apiUrl}/nfts/${tokenId}/transactions?status=success&function=addOffer&size=1&receiver=${_chainMeta?.contracts?.market}`).then((res) => {
       const txs = res.data;
       if (txs.length > 0) {
         const tx = txs[0];
@@ -77,7 +93,7 @@ export default function DataNFTDetails(props: DataNFTDetailsProps) {
         isClosable: true,
       });
     });
-  }, []);
+  }
 
   function decodeNftAttributes(nft: any) {
     const json = JSON.parse(JSON.stringify(jsonData));
@@ -111,26 +127,30 @@ export default function DataNFTDetails(props: DataNFTDetailsProps) {
   return (
     <Box>{!isLoadingNftData() ? <Box>
       <Flex direction={"column"} alignItems={"flex-start"}>
-        <Heading size="lg" marginBottom={4}>
-          Data NFT Marketplace
-        </Heading>
-        <HStack>
-          <Button
-            colorScheme="teal"
-            width={{ base: "120px", md: "160px" }}
-            _disabled={{ opacity: 1 }}
-            fontSize={{ base: "sm", md: "md" }}
-            onClick={() => {
-              navigate("/datanfts/marketplace/market/0");
-            }}
-            marginRight={2}>
-            Public Marketplace
-          </Button>
-          <Link href={nftExplorerUrl} isExternal>
-            {nftData.name} <ExternalLinkIcon mx="2px" />
-          </Link>
-        </HStack>
-        <Box width={"100%"} marginY={"56px"}>
+        {tokenIdParam &&
+          <>
+            <Heading size="lg" marginBottom={4}>
+              Data NFT Marketplace
+            </Heading>
+            <HStack>
+              <Button
+                colorScheme="teal"
+                width={{ base: "120px", md: "160px" }}
+                _disabled={{ opacity: 1 }}
+                fontSize={{ base: "sm", md: "md" }}
+                onClick={() => {
+                  navigate("/datanfts/marketplace/market/0");
+                }}
+                marginRight={2}>
+                Public Marketplace
+              </Button>
+              <Link href={nftExplorerUrl} isExternal>
+                {nftData.name} <ExternalLinkIcon mx="2px" />
+              </Link>
+            </HStack>
+          </>
+        }
+        <Box width={"100%"} marginY={tokenIdParam ? "56px" : "30px"}>
           <Stack
             flexDirection={{ base: "column", md: "row" }}
             justifyContent={{ base: "center", md: "flex-start" }}
