+++ src/Launch/LandingPage.tsx
@@ -1,19 +1,9 @@
-import React, { useState, useEffect } from "react";
+import React, { useEffect } from "react";
 import {
   Box,
-  Button,
-  Container,
   Flex,
   Heading,
-  HStack,
   Image,
-  Popover,
-  PopoverArrow,
-  PopoverBody,
-  PopoverCloseButton,
-  PopoverContent,
-  PopoverHeader,
-  PopoverTrigger,
   Text,
   Center,
   Link,
@@ -25,24 +15,28 @@ import {
 } from "@chakra-ui/react";
 import imgHeroDataNFTs from "img/landing/hero-data-nfts.png";
 import imgHeroMetaverseMask from "img/landing/hero-metaverse-mask.png";
-import logoSmlD from "img/logo-sml-d.png";
-import logoSmlL from "img/logo-sml-l.png";
-import AppFooter from "Sections/AppFooter";
+import { styleStrings } from "libs/util";
 import RecentDataNFTs from "Sections/RecentDataNFTs";
+// import { useChainMeta } from "store/ChainMetaContext";
 
-const LandingPage = ({ onLaunchMode }: { onLaunchMode?: any }) => {
+const LandingPage = () => {
   const { colorMode } = useColorMode();
+  // const { chainMeta: _chainMeta } = useChainMeta();
+
+  // useEffect(() => {
+  //   console.log('********** AppHeader LOAD _chainMeta ', _chainMeta);
+  // }, []);
 
   let containerShadow = "rgb(255 255 255 / 16%) 0px 10px 36px 0px, rgb(255 255 255 / 6%) 0px 0px 0px 1px";
-  let gradientBorder = "linear-gradient(black, black) padding-box, linear-gradient(to right, #FF439D, #00C797) border-box";
+  let gradientBorder = styleStrings.gradientBorderMulticolor;
 
   if (colorMode === "light") {
     containerShadow = "rgb(0 0 0 / 16%) 0px 10px 36px 0px, rgb(0 0 0 / 6%) 0px 0px 0px 1px";
-    gradientBorder = "linear-gradient(white, white) padding-box, linear-gradient(to right, #FF439D, #00C797) border-box";
+    gradientBorder = styleStrings.gradientBorderMulticolorLight;
   }
 
   return (
-    <Container maxW="container.xl">
+    <Box>
       <Flex
         bgColor={colorMode === "dark" ? "black" : "white"}
         flexDirection="column"
@@ -50,25 +44,6 @@ const LandingPage = ({ onLaunchMode }: { onLaunchMode?: any }) => {
         minH="100vh"
         boxShadow={containerShadow}
         zIndex={2}>
-
-        <Flex
-          h="5rem"
-          alignItems="center"
-          justifyContent="space-between"
-          backgroundColor={colorMode === "light" ? "white" : "black"}
-          borderBottom="solid .1rem"
-          borderColor="teal.300"
-          p="5">
-          <HStack alignItems={"center"} spacing={4}>
-            <Image boxSize="48px" height="auto" src={colorMode === "light" ? logoSmlL : logoSmlD} alt="Itheum Data DEX" />
-
-            <Heading size={"md"} display={["none", "block"]}>
-              Itheum Data DEX
-            </Heading>
-          </HStack>
-
-          {onLaunchMode && <PopupChainSelectorForWallet onMxEnvPick={onLaunchMode} />}
-        </Flex>
 
         <Box backgroundColor={colorMode === "light" ? "white" : "black"} flexGrow="1">
           <Flex w="100%"
@@ -108,7 +83,7 @@ const LandingPage = ({ onLaunchMode }: { onLaunchMode?: any }) => {
           </Flex>
 
           <Box backgroundColor="none" w="95%" m="auto" pt="10" pb="10">
-            <RecentDataNFTs headingText="Recent Data NFTs" networkId={"ED"} />
+            <RecentDataNFTs headingText="Recent Data NFTs" networkId={"ED"} borderMultiColorStyle={true} />
           </Box>
 
           <Box backgroundColor="none" w="95%" m="auto" pt="10" pb="10">
@@ -182,62 +157,8 @@ const LandingPage = ({ onLaunchMode }: { onLaunchMode?: any }) => {
           </Box>
 
         </Box>
-
-        <AppFooter />
       </Flex>
-    </Container>
-  );
-};
-
-const PopupChainSelectorForWallet = ({ onMxEnvPick }: { onMxEnvPick: any }) => {
-  const [showMxEnvPicker, setShowMxEnvPicker] = useState(false);
-
-  return (
-    <Popover
-      isOpen={showMxEnvPicker}
-      onOpen={() => setShowMxEnvPicker(true)}
-      onClose={() => setShowMxEnvPicker(false)}
-      closeOnBlur={true}
-      isLazy
-      lazyBehavior="keepMounted">
-      <HStack>
-        <PopoverTrigger>
-          <Button colorScheme="teal" fontSize={{ base: "sm", md: "md" }}>
-            Connect MultiversX Wallet
-          </Button>
-        </PopoverTrigger>
-      </HStack>
-
-      <PopoverContent>
-        <PopoverArrow />
-        <PopoverCloseButton />
-        <PopoverHeader>
-          <Text fontSize="md">Please pick a MultiversX environment</Text>
-        </PopoverHeader>
-        <PopoverBody>
-          <Button
-            size="sm"
-            onClick={() => {
-              setShowMxEnvPicker(false);
-              onMxEnvPick("mx", "mainnet");
-            }}>
-            {" "}
-            Mainnet
-          </Button>
-
-          <Button
-            size="sm"
-            ml="2"
-            onClick={() => {
-              setShowMxEnvPicker(false);
-              onMxEnvPick("mx", "devnet");
-            }}>
-            {" "}
-            Devnet
-          </Button>
-        </PopoverBody>
-      </PopoverContent>
-    </Popover>
+    </Box>
   );
 };
