+++ src/libs/util.ts
@@ -198,6 +198,8 @@ export const MENU = {
   TRUSTEDCOMP: 14,
   FAUCET: 14,
   CLAIMS: 15,
+  LANDING: 16,
+  NFTDETAILS: 17,
 };
 
 export const PATHS = {
@@ -216,6 +218,8 @@ export const PATHS = {
   datacoalitions: [12, [2]],
   personaldataproof: [13, [0]],
   trustedcomputation: [14, [4]],
+  nftdetails: [17, [4]],
+  offer: [17, [4]],
 };
 
 export const CHAINS = {
@@ -507,4 +511,11 @@ export const dataCATDemoUserData = {
       },
     },
   },
+};
+
+export const styleStrings = {
+  gradientBorderMulticolor: 'linear-gradient(black, black) padding-box, linear-gradient(to right, #FF439D, rgb(79 209 197 / 100%)) border-box',
+  gradientBorderMulticolorLight: 'linear-gradient(black, black) padding-box, linear-gradient(to right, #FF439D, rgb(79 209 197 / 100%)) border-box',
+  gradientBorderPassive: 'linear-gradient(black, black) padding-box, linear-gradient(to right, rgb(79 209 197 / 20%), rgb(79 209 197 / 60%)) border-box',
+  gradientBorderPassiveLight: 'linear-gradient(white, white) padding-box, linear-gradient(to right, rgb(79 209 197 / 20%), rgb(79 209 197 / 60%)) border-box',
 };
